requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.28.1

# LLM Providers
google-generativeai==0.8.4
groq==0.15.0
openai==1.59.7

# Embeddings
sentence-transformers==3.3.1
torch==2.5.1

# Database
aiosqlite==0.20.0

# Cache
redis==5.2.1

# Statistics
numpy==2.2.1
scipy==1.15.0

# Monitoring
prometheus-client==0.21.1

# Utilities
structlog==24.4.0
aiohttp==3.11.11

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Benchmarks
locust==2.32.6
